services/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# Azure SDKs
azure-identity==1.19.0
azure-cosmos==4.8.0
azure-storage-blob==12.23.1

# Microsoft Agent Framework (preview)
# Install separately with: pip install --pre agent-framework-azure-ai
# agent-framework-azure-ai

# Utilities
httpx==0.27.2
python-json-logger==3.1.0
